+++ tests/fetch/test_fetch_ubuntu.py
@@ -2,7 +2,6 @@ import six
 import subprocess
 import io
 import os
-import tempfile
 
 from tests.helpers import patch_open
 from testtools import TestCase
@@ -11,7 +10,6 @@ from mock import (
     MagicMock,
     call,
     sentinel,
-    ANY,
 )
 from charmhelpers.fetch import ubuntu as fetch
 
@@ -20,13 +18,8 @@ if six.PY3:
 else:
     builtin_open = '__builtin__.open'
 
-# mocked return of openstack.lsb_release()
-FAKE_RELEASE = {
-    'DISTRIB_CODENAME': 'precise',
-    'DISTRIB_RELEASE': '12.04',
-    'DISTRIB_ID': 'Ubuntu',
-    'DISTRIB_DESCRIPTION': '"Ubuntu 12.04"'
-}
+# mocked return of openstack.get_distrib_codename()
+FAKE_CODENAME = 'precise'
 
 url = 'deb ' + fetch.CLOUD_ARCHIVE_URL
 UCA_SOURCES = [
@@ -57,8 +50,14 @@ CQoLBBYCAwECHgECF4AACgkQimhEop9oEE7kJAP/eTBgq3Mhbvo0d8elMOuqZx3nmU7gSyPh
 ep0zYIRZ5TJWl/7PRtvp0CJA6N6ZywYTQ/4ANHhpibcHZkh8K0AzUvsGXnJRSFoJeqyDbD91
 EhoO+4ZfHs2HvRBQEDZILMa2OyuB497E5Mmyua3HDEOrG2cVLllsUZzpTFCx8NgeMHk=
 =jLBm
------END PGP PUBLIC KEY BLOCK-----
-"""
+-----END PGP PUBLIC KEY BLOCK-----"""
+
+PGP_KEY_BIN_PGP = b'\x98\x8d\x04P!2L\x01\x04\x00\xcb\x94\xc7\'\xe2z\x00\x82\xc2~\t\xfd\xcd\'\xc3\x93\xd4M"]\x9f\x91j\x89D\xea\xea#\x1f\x8a>\xb9\xc5\xfdo\x97H?3\x1a\xb0\xd2\x07\xe0\x00\x92,\x08;\xce\xf4\xdf]\x96\x82\xc0\xc0^\x85P\x13 \xe4\xcc\xbb[(Q=0\n\x17\x9d9L\xa2[\xa3%\xf7>\x9e\t\xd5\xf55-_\x10\x15S\xa9\xf3\xbb\xd4\x87Ts\x07\xa1\x94\xed\xc4\xa0\x96\x19\xdb/\xce`\x99\xb2S\x96Y3\x91\xbd\xe1\xe1\x8c;[\xe7\x9d\x8d\xc4?\x00\x11\x01\x00\x01\xb4+Launchpad PPA for Ubuntu Cloud Archive Team\x88\xb8\x04\x13\x01\x02\x00"\x05\x02P!2L\x02\x1b\x03\x06\x0b\t\x08\x07\x03\x02\x06\x15\x08\x02\t\n\x0b\x04\x16\x02\x03\x01\x02\x1e\x01\x02\x17\x80\x00\n\t\x10\x8ahD\xa2\x9fh\x10N\xe4$\x03\xffy0`\xabs!n\xfa4w\xc7\xa50\xeb\xaag\x1d\xe7\x99N\xe0K#\xe1z\x9d3`\x84Y\xe52V\x97\xfe\xcfF\xdb\xe9\xd0"@\xe8\xde\x99\xcb\x06\x13C\xfe\x004xi\x89\xb7\x07fH|+@3R\xfb\x06^rQHZ\tz\xac\x83l?u\x12\x1a\x0e\xfb\x86_\x1e\xcd\x87\xbd\x10P\x106H,\xc6\xb6;+\x81\xe3\xde\xc4\xe4\xc9\xb2\xb9\xad\xc7\x0cC\xab\x1bg\x15.YlQ\x9c\xe9LP\xb1\xf0\xd8\x1e0y' # noqa
+
+# a keyid can be retrieved by the ASCII armor-encoded key using this:
+# cat testkey.asc | gpg --with-colons --import-options import-show --dry-run
+# --import
+PGP_KEY_ID = '8a6844a29f68104e'
 
 FAKE_APT_CACHE = {
     # an installed package
@@ -128,40 +127,138 @@ class FetchTest(TestCase):
                          fetch.filter_missing_packages(['pkga', 'pkgb']))
 
     @patch.object(fetch, 'log', lambda *args, **kwargs: None)
-    def test_import_apt_key_radix(self):
-        """Ensure shell out apt-key during key import"""
-        with patch('subprocess.check_call') as _subp:
-            fetch.import_key('foo')
-            cmd = ['apt-key', 'adv', '--keyserver',
-                   'hkp://keyserver.ubuntu.com:80', '--recv-keys', 'foo']
-            _subp.assert_called_with(cmd)
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
+    def test_import_apt_key_radix(self, dearmor_gpg_key,
+                                  w_keyfile):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        with patch('subprocess.check_output') as _subp_check_output:
+            curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                                 '/pks/lookup?op=get&options=mr'
+                                 '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+            def check_output_side_effect(command, env):
+                return {
+                    ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+                }[' '.join(command)]
+            _subp_check_output.side_effect = check_output_side_effect
+
+            fetch.import_key(PGP_KEY_ID)
+            _subp_check_output.assert_called_with(curl_cmd, env=None)
+        w_keyfile.assert_called_once_with(key_name=PGP_KEY_ID,
+                                          key_material=PGP_KEY_BIN_PGP)
 
     @patch.object(fetch, 'log', lambda *args, **kwargs: None)
-    def test_import_apt_key_ascii_armor(self):
-        with tempfile.NamedTemporaryFile() as tmp:
-            with patch.object(fetch, 'NamedTemporaryFile') as mock_tmpfile:
-                tmpfile = mock_tmpfile.return_value
-                tmpfile.__enter__.return_value = tmpfile
-                tmpfile.name = tmp.name
-                with patch('subprocess.check_call') as _subp:
-                    fetch.import_key(PGP_KEY_ASCII_ARMOR)
-                    cmd = ['apt-key', 'add', tmp.name]
-                    _subp.assert_called_with(cmd)
-                with open(tmp.name, 'r') as f:
-                    self.assertEqual(PGP_KEY_ASCII_ARMOR, f.read())
+    @patch.object(os, 'getenv')
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
+    def test_import_apt_key_radix_https_proxy(self, dearmor_gpg_key,
+                                              w_keyfile, getenv):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        def get_env_side_effect(var):
+            return {
+                'HTTPS_PROXY': 'http://squid.internal:3128',
+                'JUJU_CHARM_HTTPS_PROXY': None,
+            }[var]
+        getenv.side_effect = get_env_side_effect
+
+        with patch('subprocess.check_output') as _subp_check_output:
+            proxy_settings = {
+                'HTTPS_PROXY': 'http://squid.internal:3128',
+                'https_proxy': 'http://squid.internal:3128',
+            }
+            curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                                 '/pks/lookup?op=get&options=mr'
+                                 '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+            def check_output_side_effect(command, env):
+                return {
+                    ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+                }[' '.join(command)]
+            _subp_check_output.side_effect = check_output_side_effect
+
+            fetch.import_key(PGP_KEY_ID)
+            _subp_check_output.assert_called_with(curl_cmd, env=proxy_settings)
+        w_keyfile.assert_called_once_with(key_name=PGP_KEY_ID,
+                                          key_material=PGP_KEY_BIN_PGP)
 
     @patch.object(fetch, 'log', lambda *args, **kwargs: None)
-    def test_import_bad_apt_key(self):
+    @patch.object(os, 'getenv')
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
+    def test_import_apt_key_radix_charm_https_proxy(self, dearmor_gpg_key,
+                                                    w_keyfile, getenv):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        def get_env_side_effect(var):
+            return {
+                'HTTPS_PROXY': None,
+                'JUJU_CHARM_HTTPS_PROXY': 'http://squid.internal:3128',
+            }[var]
+        getenv.side_effect = get_env_side_effect
+
+        with patch('subprocess.check_output') as _subp_check_output:
+            proxy_settings = {
+                'HTTPS_PROXY': 'http://squid.internal:3128',
+                'https_proxy': 'http://squid.internal:3128',
+            }
+            curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                                 '/pks/lookup?op=get&options=mr'
+                                 '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+            def check_output_side_effect(command, env):
+                return {
+                    ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+                }[' '.join(command)]
+            _subp_check_output.side_effect = check_output_side_effect
+
+            fetch.import_key(PGP_KEY_ID)
+            _subp_check_output.assert_called_with(curl_cmd, env=proxy_settings)
+        w_keyfile.assert_called_once_with(key_name=PGP_KEY_ID,
+                                          key_material=PGP_KEY_BIN_PGP)
+
+    @patch.object(fetch, 'log', lambda *args, **kwargs: None)
+    @patch.object(fetch, '_dearmor_gpg_key')
+    @patch('subprocess.check_output')
+    def test_import_bad_apt_key(self, check_output, dearmor_gpg_key):
         """Ensure error when importing apt key fails"""
-        with patch('subprocess.check_call') as _subp:
-            cmd = ['apt-key', 'adv', '--keyserver',
-                   'hkp://keyserver.ubuntu.com:80', '--recv-keys', 'foo']
-            _subp.side_effect = subprocess.CalledProcessError(1, cmd, '')
-            try:
-                fetch.import_key('foo')
-                assert False
-            except fetch.GPGKeyError as e:
-                self.assertEqual(str(e), "Error importing PGP key 'foo'")
+        errmsg = ('Invalid GPG key material. Check your network setup'
+                  ' (MTU, routing, DNS) and/or proxy server settings'
+                  ' as well as destination keyserver status.')
+        bad_keyid = 'foo'
+
+        curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                             '/pks/lookup?op=get&options=mr'
+                             '&exact=on&search=0x{}').format(bad_keyid)]
+
+        def check_output_side_effect(command, env):
+            return {
+                ' '.join(curl_cmd): 'foobar',
+            }[' '.join(command)]
+        check_output.side_effect = check_output_side_effect
+
+        def dearmor_side_effect(key_asc):
+            raise fetch.GPGKeyError(errmsg)
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+        try:
+            fetch.import_key(bad_keyid)
+            assert False
+        except fetch.GPGKeyError as e:
+            self.assertEqual(str(e), errmsg)
 
     @patch('charmhelpers.fetch.ubuntu.log')
     def test_add_source_none_ubuntu(self, log):
@@ -238,46 +335,160 @@ class FetchTest(TestCase):
         check_call.assert_called_with(
             ['add-apt-repository', '--yes', source])
 
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
     @patch('charmhelpers.fetch.ubuntu.log')
+    @patch('subprocess.check_output')
     @patch('subprocess.check_call')
-    def test_add_source_http_and_key_id(self, check_call, log):
+    def test_add_source_http_and_key_id(self, check_call, check_output, log,
+                                        dearmor_gpg_key,
+                                        w_keyfile):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                             '/pks/lookup?op=get&options=mr'
+                             '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+        def check_output_side_effect(command, env):
+            return {
+                ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+            }[' '.join(command)]
+        check_output.side_effect = check_output_side_effect
         source = "http://archive.ubuntu.com/ubuntu raring-backports main"
-        key_id = "akey"
         check_call.return_value = 0  # Successful exit code
-        fetch.add_source(source=source, key=key_id)
-        check_call.assert_has_calls([
-            call(['add-apt-repository', '--yes', source]),
-            call(['apt-key', 'adv', '--keyserver',
-                  'hkp://keyserver.ubuntu.com:80', '--recv-keys', key_id])
+        fetch.add_source(source=source, key=PGP_KEY_ID)
+        check_call.assert_any_call(['add-apt-repository', '--yes', source]),
+        check_output.assert_has_calls([
+            call(['curl', ('https://keyserver.ubuntu.com'
+                           '/pks/lookup?op=get&options=mr'
+                           '&exact=on&search=0x{}').format(PGP_KEY_ID)],
+                 env=None),
         ])
 
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
     @patch('charmhelpers.fetch.ubuntu.log')
+    @patch('subprocess.check_output')
     @patch('subprocess.check_call')
-    def test_add_source_https_and_key_id(self, check_call, log):
+    def test_add_source_https_and_key_id(self, check_call, check_output, log,
+                                         dearmor_gpg_key,
+                                         w_keyfile):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                             '/pks/lookup?op=get&options=mr'
+                             '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+        def check_output_side_effect(command, env):
+            return {
+                ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+            }[' '.join(command)]
+        check_output.side_effect = check_output_side_effect
+
+        check_call.return_value = 0
+
         source = "https://USER:PASS@private-ppa.launchpad.net/project/awesome"
-        key_id = "GPGPGP"
-        check_call.return_value = 0  # Success from both calls
-        fetch.add_source(source=source, key=key_id)
-        check_call.assert_has_calls([
-            call(['add-apt-repository', '--yes', source]),
-            call(['apt-key', 'adv', '--keyserver',
-                  'hkp://keyserver.ubuntu.com:80', '--recv-keys', key_id])
+        fetch.add_source(source=source, key=PGP_KEY_ID)
+        check_call.assert_any_call(['add-apt-repository', '--yes', source]),
+        check_output.assert_has_calls([
+            call(['curl', ('https://keyserver.ubuntu.com'
+                           '/pks/lookup?op=get&options=mr'
+                           '&exact=on&search=0x{}').format(PGP_KEY_ID)],
+                 env=None),
         ])
 
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
     @patch('charmhelpers.fetch.ubuntu.log')
+    @patch.object(fetch, 'get_distrib_codename')
     @patch('subprocess.check_call')
-    def test_add_source_http_and_key(self, check_call, log):
+    @patch('subprocess.Popen')
+    def test_add_source_http_and_key_gpg1(self, popen, check_call,
+                                          get_distrib_codename, log,
+                                          dearmor_gpg_key,
+                                          w_keyfile):
         source = "http://archive.ubuntu.com/ubuntu raring-backports main"
-        key = '''
-            -----BEGIN PGP PUBLIC KEY BLOCK-----
-            [...]
-            -----END PGP PUBLIC KEY BLOCK-----
-            '''
-        with patch('subprocess.check_call') as check_call:
-            check_call.return_value = 0
-            fetch.add_source(source=source, key=key)
-            check_call.assert_any_call(['add-apt-repository', '--yes', source])
-            check_call.assert_any_call(['apt-key', 'add', ANY])
+        key = PGP_KEY_ASCII_ARMOR
+        key_bytes = PGP_KEY_ASCII_ARMOR.encode('utf-8')
+        get_distrib_codename.return_value = 'trusty'
+        check_call.return_value = 0
+
+        expected_key = '35F77D63B5CEC106C577ED856E85A86E4652B4E6'
+        if six.PY3:
+            popen.return_value.communicate.return_value = [b"""
+pub:-:1024:1:6E85A86E4652B4E6:2009-01-18:::-:Launchpad PPA for Landscape:
+fpr:::::::::35F77D63B5CEC106C577ED856E85A86E4652B4E6:
+            """, b'']
+        else:
+            popen.return_value.communicate.return_value = ["""
+pub:-:1024:1:6E85A86E4652B4E6:2009-01-18:::-:Launchpad PPA for Landscape:
+fpr:::::::::35F77D63B5CEC106C577ED856E85A86E4652B4E6:
+            """, '']
+
+        dearmor_gpg_key.return_value = PGP_KEY_BIN_PGP
+
+        fetch.add_source(source=source, key=key)
+        popen.assert_called_with(
+            ['gpg', '--with-colons', '--with-fingerprint'],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            stdin=subprocess.PIPE)
+        dearmor_gpg_key.assert_called_with(key_bytes)
+        w_keyfile.assert_called_with(key_name=expected_key,
+                                     key_material=PGP_KEY_BIN_PGP)
+        check_call.assert_any_call(['add-apt-repository', '--yes', source]),
+
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
+    @patch('charmhelpers.fetch.ubuntu.log')
+    @patch.object(fetch, 'get_distrib_codename')
+    @patch('subprocess.check_call')
+    @patch('subprocess.Popen')
+    def test_add_source_http_and_key_gpg2(self, popen, check_call,
+                                          get_distrib_codename, log,
+                                          dearmor_gpg_key,
+                                          w_keyfile):
+        source = "http://archive.ubuntu.com/ubuntu raring-backports main"
+        key = PGP_KEY_ASCII_ARMOR
+        key_bytes = PGP_KEY_ASCII_ARMOR.encode('utf-8')
+        get_distrib_codename.return_value = 'bionic'
+        check_call.return_value = 0
+
+        expected_key = '35F77D63B5CEC106C577ED856E85A86E4652B4E6'
+
+        if six.PY3:
+            popen.return_value.communicate.return_value = [b"""
+fpr:::::::::35F77D63B5CEC106C577ED856E85A86E4652B4E6:
+uid:-::::1232306042::52FE92E6867B4C099AA1A1877A804A965F41A98C::ppa::::::::::0:
+            """, b'']
+        else:
+            # python2 on a distro with gpg2 (unlikely, but possible)
+            popen.return_value.communicate.return_value = ["""
+fpr:::::::::35F77D63B5CEC106C577ED856E85A86E4652B4E6:
+uid:-::::1232306042::52FE92E6867B4C099AA1A1877A804A965F41A98C::ppa::::::::::0:
+            """, '']
+
+        dearmor_gpg_key.return_value = PGP_KEY_BIN_PGP
+
+        fetch.add_source(source=source, key=key)
+        popen.assert_called_with(
+            ['gpg', '--with-colons', '--import-options', 'show-only',
+             '--import', '--dry-run'],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            stdin=subprocess.PIPE)
+        dearmor_gpg_key.assert_called_with(key_bytes)
+        w_keyfile.assert_called_with(key_name=expected_key,
+                                     key_material=PGP_KEY_BIN_PGP)
+        check_call.assert_any_call(['add-apt-repository', '--yes', source]),
 
     def test_add_source_cloud_invalid_pocket(self):
         source = "cloud:havana-updates"
@@ -287,11 +498,12 @@ class FetchTest(TestCase):
     @patch('charmhelpers.fetch.ubuntu.log')
     @patch.object(fetch, 'filter_installed_packages')
     @patch.object(fetch, 'apt_install')
-    @patch.object(fetch, 'lsb_release')
-    def test_add_source_cloud_pocket_style(self, lsb_release, apt_install,
+    @patch.object(fetch, 'get_distrib_codename')
+    def test_add_source_cloud_pocket_style(self, get_distrib_codename,
+                                           apt_install,
                                            filter_pkg, log):
         source = "cloud:precise-updates/havana"
-        lsb_release.return_value = {'DISTRIB_CODENAME': 'precise'}
+        get_distrib_codename.return_value = 'precise'
         result = ('# Ubuntu Cloud Archive\n'
                   'deb http://ubuntu-cloud.archive.canonical.com/ubuntu'
                   ' precise-updates/havana main\n')
@@ -304,11 +516,11 @@ class FetchTest(TestCase):
     @patch('charmhelpers.fetch.ubuntu.log')
     @patch.object(fetch, 'filter_installed_packages')
     @patch.object(fetch, 'apt_install')
-    @patch.object(fetch, 'lsb_release')
-    def test_add_source_cloud_os_style(self, lsb_release, apt_install,
+    @patch.object(fetch, 'get_distrib_codename')
+    def test_add_source_cloud_os_style(self, get_distrib_codename, apt_install,
                                        filter_pkg, log):
         source = "cloud:precise-havana"
-        lsb_release.return_value = {'DISTRIB_CODENAME': 'precise'}
+        get_distrib_codename.return_value = 'precise'
         result = ('# Ubuntu Cloud Archive\n'
                   'deb http://ubuntu-cloud.archive.canonical.com/ubuntu'
                   ' precise-updates/havana main\n')
@@ -332,68 +544,105 @@ class FetchTest(TestCase):
         filter_pkg.assert_called_with(['ubuntu-cloud-keyring'])
 
     @patch('charmhelpers.fetch.ubuntu.log')
-    @patch.object(fetch, 'lsb_release')
+    @patch.object(fetch, 'get_distrib_codename')
     @patch('platform.machine')
-    def test_add_source_proposed_x86_64(self, _machine, lsb_release, log):
+    def test_add_source_proposed_x86_64(self, _machine,
+                                        get_distrib_codename, log):
         source = "proposed"
         result = ('# Proposed\n'
                   'deb http://archive.ubuntu.com/ubuntu precise-proposed'
                   ' main universe multiverse restricted\n')
-        lsb_release.return_value = {'DISTRIB_CODENAME': 'precise'}
+        get_distrib_codename.return_value = 'precise'
         _machine.return_value = 'x86_64'
         with patch_open() as (mock_open, mock_file):
             fetch.add_source(source=source)
             mock_file.write.assert_called_with(result)
 
     @patch('charmhelpers.fetch.ubuntu.log')
-    @patch.object(fetch, 'lsb_release')
+    @patch.object(fetch, 'get_distrib_codename')
     @patch('platform.machine')
-    def test_add_source_proposed_ppc64le(self, _machine, lsb_release, log):
+    def test_add_source_proposed_ppc64le(self, _machine,
+                                         get_distrib_codename, log):
         source = "proposed"
         result = (
             "# Proposed\n"
             "deb http://ports.ubuntu.com/ubuntu-ports precise-proposed main "
             "universe multiverse restricted\n")
-        lsb_release.return_value = {'DISTRIB_CODENAME': 'precise'}
+        get_distrib_codename.return_value = 'precise'
         _machine.return_value = 'ppc64le'
         with patch_open() as (mock_open, mock_file):
             fetch.add_source(source=source)
             mock_file.write.assert_called_with(result)
 
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
     @patch('charmhelpers.fetch.ubuntu.log')
+    @patch('subprocess.check_output')
     @patch('subprocess.check_call')
-    def test_add_source_http_and_key_id_ubuntu(self, check_call, log):
+    def test_add_source_http_and_key_id_ubuntu(self, check_call, check_output,
+                                               log, dearmor_gpg_key,
+                                               w_keyfile):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                             '/pks/lookup?op=get&options=mr'
+                             '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+        def check_output_side_effect(command, env):
+            return {
+                ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+            }[' '.join(command)]
+        check_output.side_effect = check_output_side_effect
+        check_call.return_value = 0
         source = "http://archive.ubuntu.com/ubuntu raring-backports main"
-        key_id = "akey"
+        key_id = PGP_KEY_ID
         fetch.add_source(source=source, key=key_id)
         check_call.assert_any_call(['add-apt-repository', '--yes', source]),
-        check_call.assert_any_call([
-            'apt-key', 'adv', '--keyserver',
-            'hkp://keyserver.ubuntu.com:80', '--recv-keys', key_id])
+        check_output.assert_has_calls([
+            call(['curl', ('https://keyserver.ubuntu.com'
+                           '/pks/lookup?op=get&options=mr'
+                           '&exact=on&search=0x{}').format(PGP_KEY_ID)],
+                 env=None),
+        ])
 
+    @patch.object(fetch, '_write_apt_gpg_keyfile')
+    @patch.object(fetch, '_dearmor_gpg_key')
     @patch('charmhelpers.fetch.ubuntu.log')
+    @patch('subprocess.check_output')
     @patch('subprocess.check_call')
-    def test_add_source_https_and_key_id_ubuntu(self, check_call, log):
+    def test_add_source_https_and_key_id_ubuntu(self, check_call, check_output,
+                                                log, dearmor_gpg_key,
+                                                w_keyfile):
+        def dearmor_side_effect(key_asc):
+            return {
+                PGP_KEY_ASCII_ARMOR: PGP_KEY_BIN_PGP,
+            }[key_asc]
+        dearmor_gpg_key.side_effect = dearmor_side_effect
+
+        curl_cmd = ['curl', ('https://keyserver.ubuntu.com'
+                             '/pks/lookup?op=get&options=mr'
+                             '&exact=on&search=0x{}').format(PGP_KEY_ID)]
+
+        def check_output_side_effect(command, env):
+            return {
+                ' '.join(curl_cmd): PGP_KEY_ASCII_ARMOR,
+            }[' '.join(command)]
+        check_output.side_effect = check_output_side_effect
+        check_call.return_value = 0
+
         source = "https://USER:PASS@private-ppa.launchpad.net/project/awesome"
-        key_id = "GPGPGP"
-        fetch.add_source(source=source, key=key_id)
+        fetch.add_source(source=source, key=PGP_KEY_ID)
         check_call.assert_any_call(['add-apt-repository', '--yes', source]),
-        check_call.assert_any_call([
-            'apt-key', 'adv', '--keyserver',
-            'hkp://keyserver.ubuntu.com:80', '--recv-keys', key_id])
-
-    @patch('charmhelpers.fetch.ubuntu.log')
-    @patch('subprocess.check_call')
-    def test_add_source_http_and_key_ubuntu(self, check_call, log):
-        source = "http://archive.ubuntu.com/ubuntu raring-backports main"
-        key = '''
-            -----BEGIN PGP PUBLIC KEY BLOCK-----
-            [...]
-            -----END PGP PUBLIC KEY BLOCK-----
-            '''
-        fetch.add_source(source=source, key=key)
-        check_call.assert_any_call(['add-apt-repository', '--yes', source])
-        check_call.assert_any_call(['apt-key', 'add', ANY])
+        check_output.assert_has_calls([
+            call(['curl', ('https://keyserver.ubuntu.com'
+                           '/pks/lookup?op=get&options=mr'
+                           '&exact=on&search=0x{}').format(PGP_KEY_ID)],
+                 env=None),
+        ])
 
     @patch('charmhelpers.fetch.ubuntu.log')
     def test_configure_bad_install_source(self, log):
@@ -403,10 +652,10 @@ class FetchTest(TestCase):
         except fetch.SourceConfigError as e:
             self.assertEqual(str(e), "Unknown source: 'foo'")
 
-    @patch('charmhelpers.fetch.ubuntu.lsb_release')
+    @patch('charmhelpers.fetch.ubuntu.get_distrib_codename')
     def test_configure_install_source_uca_staging(self, _lsb):
         """Test configuring installation source from UCA staging sources"""
-        _lsb.return_value = FAKE_RELEASE
+        _lsb.return_value = FAKE_CODENAME
         # staging pockets are configured as PPAs
         with patch('subprocess.check_call') as _subp:
             src = 'cloud:precise-folsom/staging'
@@ -417,12 +666,12 @@ class FetchTest(TestCase):
 
     @patch(builtin_open)
     @patch('charmhelpers.fetch.ubuntu.apt_install')
-    @patch('charmhelpers.fetch.ubuntu.lsb_release')
+    @patch('charmhelpers.fetch.ubuntu.get_distrib_codename')
     @patch('charmhelpers.fetch.ubuntu.filter_installed_packages')
     def test_configure_install_source_uca_repos(
             self, _fip, _lsb, _install, _open):
         """Test configuring installation source from UCA sources"""
-        _lsb.return_value = FAKE_RELEASE
+        _lsb.return_value = FAKE_CODENAME
         _file = MagicMock(spec=io.FileIO)
         _open.return_value = _file
         _fip.side_effect = lambda x: x